Move InitContainer function to its own file
